refactor(seeders): consolidate seed logic into unified seeders/ package
Move skills/seeder.py → seeders/skill_seeder.py
Move services/seed.py   → seeders/provider_seeder.py

All seeder imports now go through cubebox.seeders, keeping the
lifespan clean and seed logic co-located in one package.
